+++ packages/@sfx/sayt-driver-plugin/CHANGELOG.md
@@ -8,5 +8,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - SFX-160: Added the `SaytDriverPlugin` module.
   - The `SaytDriverPlugin` module acts as a mediator between the Logic and View layers for prebuilt modules.
-  - SFX-295: Added the associated group ID to response events.
+  - SFX-248: The `@sfx/events` package is used for event names and payload interfaces.
   - SFX-333: Added retrieval and event emission of products.